fix(issue-633): migrate to http handler
