+++ services/machine/endpoint/endpointpb/accelerator.proto
