+++ services/machine/endpoint/endpointpb/cpu.proto
