+++ services/machine/endpoint/endpointpb/machine.proto
