+++ services/machine/endpoint/endpointpb/memory_module.proto
