+++ services/machine/endpoint/endpointpb/nic.proto
