+++ services/machine/endpoint/endpointpb/register_machine_request.proto
