+++ services/machine/endpoint/endpointpb/register_machine_response.proto
