+++ services/machine/endpoint/endpointpb/register_machine_service.proto
